requirements.txt
opencv-python==4.11.0.86
Flask==3.1.0
numpy==2.0.2
pillow==11.1.0
tensorflow==2.18.0